server/requirements.txt
alembic==1.13.1
async-timeout==4.0.3
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==2.3.8
pycryptodome
Flask-WTF
